Add files via upload
"Explore the impact of COVID-19 with this SQL project. We've delved into the data to uncover trends, patterns, and critical insights. Discover the power of structured query language in unraveling the pandemic's story. From cases and testing to vaccinations and more, this project offers a deep dive into the COVID-19 dataset to help you understand the evolving landscape."